+++ service/fixed/databases/service.go
@@ -85,6 +85,19 @@ func (a *API) Update(ctx context.Context, subscription int, database int, update
 	return a.taskWaiter.Wait(ctx, *task.ID)
 }
 
+// UpgradeRedisVersion will upgrade the Redis version of an existing fixed database.
+func (a *API) UpgradeRedisVersion(ctx context.Context, subscription int, database int, upgradeVersion UpgradeRedisVersion) error {
+	var task internal.TaskResponse
+	err := a.client.Post(ctx, fmt.Sprintf("upgrade fixed database %d version for subscription %d", database, subscription), fmt.Sprintf("/fixed/subscriptions/%d/databases/%d/upgrade", subscription, database), upgradeVersion, &task)
+	if err != nil {
+		return err
+	}
+
+	a.logger.Printf("Waiting for fixed database %d for subscription %d to finish being upgraded", database, subscription)
+
+	return a.taskWaiter.Wait(ctx, *task.ID)
+}
+
 // Delete will destroy an existing fixed database.
 func (a *API) Delete(ctx context.Context, subscription int, database int) error {
 	var task internal.TaskResponse