+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 * Added `ID` field to `Region` struct in account model
+* Added Redis version support for Essentials databases: `RedisVersion` field and `UpgradeRedisVersion()` method
 
 ## 0.41.0 (3rd November 2025)